requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
PyAudio==0.2.14
razorpay==1.4.2
requests==2.32.3
SpeechRecognition==3.10.4
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
